Add token table to flex scanner.
Add a symbol/token table to the flex/bison grammar so that we can account for
all of memory if the parser FSM barfs.  This is because we have token strings on
the FSM stack that don't get fed to config_{block|stmt|term} until the reduce
happens.  If we error out of the rule instead, these tokens and their memory go poof!
The table keeps track of things and everything gets freed when the parser is released.

An added benefit is that we search first which saves for us because almost everything
is pre-defined tokens and little is free form text or random declarationsd...

Signed-off-by: Jim Lieb <lieb@sea-troll.net>